US23 Cadastro Especialidade Médica

+++ isaude-core/src/main/java/br/com/its/isaude/core/generic/impl/dao/GenericHibernateDAO.java
@@ -30,7 +30,7 @@ public abstract class GenericHibernateDAO<T> implements GenericDAO<T> {
 	}
 
 
-	public T getById(Long id) {
+	public T getById(Integer id) {
 		return (T) getCurrentSession().get(persistentClass, id);
 	}
